pyproject.toml
[project]
name = "agora-library-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "agora_library_server", from = "src"}]

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.11.10"
lint = "^1.2.1"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev agora_library_server/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=agora_library_server -vv'
post_test = 'coverage html'
